Clarify and remove inconsistencies
There were several steps the either were unclear or were inconsistent
with other steps. The largest inconsistency was in regard to whether the
student's code should be written procedurally or OO. E.g. It says to make
a NMF class in the main part, but then says to convert your NMF code into a
class as part of the "Extra" section.